🚚 Move FlameShock to shared folder
