+++ src/parser/shaman/elemental/CombatLogParser.tsx
@@ -22,7 +22,6 @@ import Icefury from 'parser/shaman/elemental/modules/talents/Icefury';
 import EchoingShock from 'parser/shaman/elemental/modules/talents/EchoingShock';
 
 import LavaSurge from 'parser/shaman/elemental/modules/core/LavaSurge';
-import FlameShock from 'parser/shaman/elemental/modules/core/FlameShock';
 import CancelledCasts from 'parser/shaman/elemental/modules/features/CancelledCasts';
 import Checklist from 'parser/shaman/elemental/modules/checklist/Module';
 
@@ -37,6 +36,7 @@ import ElementalBlast from '../shared/talents/ElementalBlast';
 import MaelstromDetails from '../shared/maelstromchart/MaelstromDetails';
 import MaelstromTracker from '../shared/maelstromchart/MaelstromTracker';
 import MaelstromTab from '../shared/maelstromchart/MaelstromTab';
+import FlameShock from '../shared/spells/FlameShock';
 
 class CombatLogParser extends CoreCombatLogParser {
   static specModules = {
